disruptor(单生产者，单消费者)

+++ disruptor/src/main/java/disruptor/MyRingBuffer.java
@@ -35,7 +35,7 @@ public class MyRingBuffer<T> {
         }
     }
 
-    public void publish(int index){
+    public void publish(Long index){
         this.singleProducerSequencer.publish(index);
         this.sequenceBarrier.signalAllWhenBlocking();
     }
